Allow dynamic arrays of motors in move commands
- Replaced StepControl methods that take a static array of motors with a
C-style functions that take a pointer to an array of motor pointers
along with a size parameter. This allows dynamic arrays to be driven,
e.g. when the desired motor sub-set is dependent on the system state
- I also added a pretty sketch remove() method that allows motors to be
removed from the array of motors currently in motion. This is useful,
for instance, if a limit-switch interrupt is encounted on one motor but
the rest should continue. This can occur in calibration senarios.