Implemented end-after-last-pulse for stm32's timers
The feature wasn't implemented and that could cause the steppers to do some extra steps before stopping. The steppers were aware of the extra steps and compensated for them on the next movement so no steps were lost but they didn't end up in the exact desired position. This commit fixes the extra steps issue.